backend/requirements.txt
# Django & Django REST Framework
django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
djangorestframework-simplejwt==5.3.0
django-filter==23.3
djoser==2.2.0
drf-yasg==1.21.7
django-phonenumber-field==7.1.0
phonenumbers==8.13.19

# PostgreSQL Adapter
psycopg2-binary==2.9.9

# PDF Processing
pypdf2==3.0.1

# AI Integration
langchain==0.0.335
openai==1.3.3

# Storage
django-storages==1.14.2
boto3==1.33.6

# Supabase
supabase==1.0.3

# Environment Variables
python-dotenv==1.0.0

# Production
gunicorn==21.2.0
whitenoise==6.5.0

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
pytest==7.4.3
pytest-django==4.7.0
